% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{bam_coverage}
\alias{bam_coverage}
\title{Calculate coverage of BAM files using bamCoverage from deepTools}
\usage{
bam_coverage(
  path_to_bamCoverage,
  bam_files,
  output_bigwigs = NULL,
  binSize = 50,
  effectiveGenomeSize,
  parallel_cores = 1
)
}
\arguments{
\item{path_to_bamCoverage}{Path to bamCoverage}

\item{bam_files}{Paths to BAM files}

\item{output_bigwigs}{Paths to save output bigWigs. Default is to replace ".bam" with ".bw".}

\item{parallel_cores}{Number of cores to use. Default is 1.}
}
\description{
Calculate coverage of BAM files using bamCoverage from deepTools
}
