% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{combineMatrices}
\alias{combineMatrices}
\title{Combine Matrices}
\usage{
combineMatrices(data.list)
}
\arguments{
\item{data.list}{List of matrices to combine}
}
\value{
Long format data.frame of combined matrices
}
\description{
Combine a list of matrices with the same number of columns into a single long
format data.frame
}

