% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{countBoxplots}
\alias{countBoxplots}
\title{Count Boxplot}
\usage{
countBoxplots(data)
}
\arguments{
\item{data}{Matrix to plot}
}
\value{
ggplot2 object containing the boxplots
}
\description{
Produce a ggplot2 object plotting boxplots from a matrix of counts
}

