% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{countDensity}
\alias{countDensity}
\title{Count Density}
\usage{
countDensity(data)
}
\arguments{
\item{data}{Matrix to plot}
}
\value{
ggplot2 object containing the density plot
}
\description{
Produce a ggplot2 object plotting densities from a matrix of counts
}

