% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{countHeatmap}
\alias{countHeatmap}
\title{Count heatmap}
\usage{
countHeatmap(data, groups = colnames(data))
}
\arguments{
\item{data}{Matrix to produce the heatmap from}
}
\value{
List counting elements of heatmap
}
\description{
Produce a clustered heatmap from a count matrix
}

