% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma_plots.R
\name{countMA}
\alias{countMA}
\title{Count MA}
\usage{
countMA(data)
}
\arguments{
\item{data}{Matrix to plot}
}
\value{
ggplot2 object containing the MA plot
}
\description{
Produce a ggplot2 object containing an MA plot from a matrix of counts
}

