% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{counts2edgeR}
\alias{counts2edgeR}
\title{Counts to edgeR}
\usage{
counts2edgeR(data, groups)
}
\arguments{
\item{data}{Matrix of counts}

\item{groups}{Vector of groups for each sample column}
}
\value{
DGEList object containing counts
}
\description{
Convert a count matrix to an edgeR DGEList object
}

