% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{counts2Objects}
\alias{counts2Objects}
\title{Counts to Objects}
\usage{
counts2Objects(data, groups, filter, methods = c("edgeR", "DESeq", "DESeq2",
  "voom"), verbose = TRUE)
}
\arguments{
\item{data}{Matrix of counts}

\item{groups}{Vector of groups for each sample column}

\item{filter}{Boolean, whether to use HTSFilter for voom}

\item{methods}{Vector of object types to convert to}

\item{verbose}{Boolean, whether to print messages showing progress}
}
\value{
List of converted objects
}
\description{
Convert a count matrix to objects required by various differential expression
testing packages
}

