% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{counts2voom}
\alias{counts2voom}
\title{Counts to voom}
\usage{
counts2voom(data, groups, filter)
}
\arguments{
\item{data}{Matrix of counts}

\item{groups}{Vector of groups for each sample column}

\item{filter}{Boolean, whether to apply HTSFilter}
}
\value{
DGEList object containing counts
}
\description{
Convert a count matrix to an edgeR DGEList object for voom
}

