% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalise.R
\name{deseq2Norm}
\alias{deseq2Norm}
\title{DESeq2 Normalise}
\usage{
deseq2Norm(count.data)
}
\arguments{
\item{count.data}{DESeqDataSet object to normalise}
}
\value{
Normalised DESeqDataSet object
}
\description{
Normalise a DESeqDataSet using DESeq2
}

