% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{deseq2Test}
\alias{deseq2Test}
\title{DESeq2 Test}
\usage{
deseq2Test(count.data, group1, group2, filter)
}
\arguments{
\item{count.data}{Normalised DESeqDataSet object to normalise}

\item{group1}{First group to test, the reference or control}

\item{group2}{Second group to test, the treatment}

\item{filter}{Boolean, whether to apply HTSFilter}
}
\value{
DESeqResults object containing test results
}
\description{
Test a normalised DESeqDataSet using DESeq2
}

