% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalise.R
\name{deseqNorm}
\alias{deseqNorm}
\title{DESeq Normalise}
\usage{
deseqNorm(count.data)
}
\arguments{
\item{count.data}{CountDataSet object to normalise}
}
\value{
Normalised CountDataSet object
}
\description{
Normalise a CountDataSet object using DESeq
}

