% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalise.R
\name{edgeRNorm}
\alias{edgeRNorm}
\title{edgeR Normalise}
\usage{
edgeRNorm(dge)
}
\arguments{
\item{dge}{DGEList object to normalise}
}
\value{
DGEList object with normalistion factors
}
\description{
Normalise a DGEList object using edgeR
}

