% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{geneSummary}
\alias{geneSummary}
\title{Gene Summary}
\usage{
geneSummary(data.list, alpha = 0.05)
}
\arguments{
\item{data.list}{List of results to summarise}

\item{alpha}{Significance level for selecting genes}
}
\value{
data.table containing summary statistics
}
\description{
Summarise the results from multiple differential expression methods by gene.
Returns a table with summary statistics and a count of the number of methods
that found the gene to be differentially expressed at a give significance
level.
}

