% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapDendro}
\alias{heatmapDendro}
\title{Heatmap Dendrogram}
\usage{
heatmapDendro(data, rows = TRUE)
}
\arguments{
\item{data}{Dendrogram data to plot}

\item{rows}{Boolean, whether the plot is for rows}
}
\value{
ggplot2 object containing the dendrogram
}
\description{
Produce a dendrogram for a heatmap using ggplot2 and ggdendro
}

