% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapLegend}
\alias{heatmapLegend}
\title{Heatmap legend}
\usage{
heatmapLegend(heatmap)
}
\arguments{
\item{heatmap}{Plot to extract the legend from}
}
\value{
The heatmap legend
}
\description{
Extract the legend from a ggplot2 heatmap
}

