% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{index_bams}
\alias{index_bams}
\title{Index BAM files}
\usage{
index_bams(path_to_samtools, bam_files, parallel_cores = 1)
}
\arguments{
\item{path_to_samtools}{Path to samtools directory}

\item{bam_files}{Paths to BAM files}

\item{parallel_cores}{Number of cores to use. Default is 1.}
}
\description{
Index BAM files
}
