% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{jaccardTable}
\alias{jaccardTable}
\title{Jaccard Table}
\usage{
jaccardTable(data.list, alpha = 0.05)
}
\arguments{
\item{data.list}{List of results to combine}

\item{alpha}{Significance level for selecting genes}
}
\value{
dataframe containing Jaccard indices
}
\description{
Produce a table of Jaccard Indices from a list of regularised differential
expression results
}

