% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto_quantify.R
\name{kallisto_quantify}
\alias{kallisto_quantify}
\title{Run kallisto on a group of FASTQ files}
\usage{
kallisto_quantify(
  path_to_kallisto,
  kallisto_index,
  forward_fastqs,
  reverse_fastqs,
  sample_names,
  output_directory,
  merged_output_prefix = "kallisto_transcript",
  messages_file = "",
  n_cores = 1,
  number_bootstraps = 100
)
}
\arguments{
\item{path_to_kallisto}{Path to kallisto executable}

\item{kallisto_index}{Path to a kallisto index}

\item{forward_fastqs}{A vector with the paths to forward FASTQ files. Each file should be the match of the corresponding file in reverse_fastqs.}

\item{reverse_fastqs}{A vector with the paths to reverse FASTQ files. Each file should be the match of the corresponding file in forward_fastqs.}

\item{sample_names}{A vector with the names of samples corresponding for each pair of samples from forward_fastqs and reverse_fastqs}

\item{output_directory}{The name of the directory to save results in. Is created if it doesn't already exist.}

\item{merged_output_prefix}{Prefix to use for names of merged output files for counts and TPM which take the form
{merged_output_prefix}_counts_merged.tsv.gz and {merged_output_prefix}_tpm_merged.tsv.gz. Default prefix is "kallisto_transcript" i.e. default output merged output files are
kallisto_transcript_counts_merged.tsv.gz and kallisto_transcript_tpm_merged.tsv.gz.}

\item{messages_file}{Name of file to save kallisto run messages. If no file name given, information is printed to stdout.}

\item{n_cores}{The number of cores to use. Default is 1.}

\item{number_bootstraps}{The number of bootstrap samples. Default is 100.}
}
\description{
Run kallisto on a group of FASTQ files
}
