% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{lengthenMatrix}
\alias{lengthenMatrix}
\title{Lengthen}
\usage{
lengthenMatrix(data)
}
\arguments{
\item{data}{Matrix to convert to long format}
}
\value{
Long format data.frame
}
\description{
Convert a matrix with informative rowanames to a long format data.frame
}

