% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{listBoxplots}
\alias{listBoxplots}
\title{List Boxplots}
\usage{
listBoxplots(data.list)
}
\arguments{
\item{data.list}{List of matrices to plot}
}
\value{
List of ggplot2 object containing boxplots
}
\description{
Plot boxplots from a list of matrices
}

