% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{listCountMA}
\alias{listCountMA}
\title{List Cout MA}
\usage{
listCountMA(data.list)
}
\arguments{
\item{data.list}{List of count matrices to plot}
}
\value{
List of ggplot2 object containing MA plots
}
\description{
Plot MA plots from a list of count matrices
}

