% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{listDensity}
\alias{listDensity}
\title{List Density}
\usage{
listDensity(data.list)
}
\arguments{
\item{data.list}{List of matrices to plot}
}
\value{
List of ggplot2 object containing histograms
}
\description{
Plot densnities from a list of matrices
}

