% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{listHeatmaps}
\alias{listHeatmaps}
\title{List Heatmaps}
\usage{
listHeatmaps(data.list, groups = colnames(data.list[[1]]))
}
\arguments{
\item{data.list}{List of matrices to plot}

\item{groups}{Vector of groups for samples}
}
\value{
List of lists containing objects for heatmaps
}
\description{
Consturct heatmaps from a list of matrices
}

