% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plots.R
\name{listMA}
\alias{listMA}
\title{List MA}
\usage{
listMA(data.list)
}
\arguments{
\item{data.list}{List of matrices to plot}
}
\value{
List of ggplot2 object containing MA plots
}
\description{
Plot MA plots from a list of matrices
}

