% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_mds.R
\name{listMDS}
\alias{listMDS}
\title{List MDS}
\usage{
listMDS(data.list, top = nrow(data.list[[1]]),
  groups = colnames(data.list[[1]]), selection = "pairwise")
}
\arguments{
\item{data.list}{List of matrices to plot}

\item{top}{Number of rows with highes variance to select for plotting}

\item{groups}{Vector of groups assigned to sample columns}

\item{selection}{Select rows in a "pariwise" manner between samples or
                 "common" across all samples}
}
\value{
List of ggplot2 objects containing MDS plots
}
\description{
Produce MDS plots from a list of matrices
}

