% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalise.R
\name{listNorm}
\alias{listNorm}
\title{List Normalise}
\usage{
listNorm(data.list, verbose = TRUE)
}
\arguments{
\item{data.list}{List of differential expression objects}

\item{verbose}{Boolean, whether to print messages showing progress}
}
\value{
List of normalised objects
}
\description{
Apply the appropriate normalisation to a list of differential expression
objects
}

