% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_mds.R
\name{listPCA}
\alias{listPCA}
\title{List PCA}
\usage{
listPCA(data.list, top = nrow(data.list[[1]]),
  groups = colnames(data.list[[1]]))
}
\arguments{
\item{data.list}{List of matrices to plot}

\item{top}{Number of rows with highes variance to select for plotting}

\item{groups}{Vector of groups assigned to sample columns}
}
\value{
List of ggplot2 objects containing PCA plots
}
\description{
Produce PCA plots from a list of matrices
}

