% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_values.R
\name{listPlotPvals}
\alias{listPlotPvals}
\title{List p-value plots}
\usage{
listPlotPvals(data.list)
}
\arguments{
\item{data.list}{List of results to plot}
}
\value{
List of ggplot2 objects containing p-value histograms
}
\description{
Plot hitograms of p-values from a list of differential expression results
}

