% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{listRegularise}
\alias{listRegularise}
\title{List Regularise}
\usage{
listRegularise(data.list)
}
\arguments{
\item{data.list}{List of differential expression results}
}
\value{
List of regularised results
}
\description{
Convert list of differential expression results to a regular format
}

