% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma_plots.R
\name{listResultsMA}
\alias{listResultsMA}
\title{List Results MA}
\usage{
listResultsMA(data.list, alpha = 0.05)
}
\arguments{
\item{data.list}{List of results to plot}

\item{alpha}{Significance level for labelling differentially
                 expressed genes}
}
\value{
List of ggplot2 object containing MA plots
}
\description{
Plot MA plots from a list of differential expression results
}

