% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{listTest}
\alias{listTest}
\title{List Test}
\usage{
listTest(data.list, group1, group2)
}
\arguments{
\item{data.list}{List of normalised differential expression objects}

\item{group1}{First group to test, the reference or control}

\item{group2}{Second group to test, the treatment}
}
\value{
List of test result objects
}
\description{
Apply the appropriate testing procedure to a list of normalised differential
expression objects
}

