% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{logCPMTransform}
\alias{logCPMTransform}
\title{Log Counts Per Million transform counts}
\usage{
logCPMTransform(data)
}
\arguments{
\item{data}{Matrix of raw counts}
}
\value{
Matrix of transformed data
}
\description{
Apply edgeR's Counts Per Million transformation to raw counts, then take the
log
}

