% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{logTransform}
\alias{logTransform}
\title{Log transform counts}
\usage{
logTransform(data)
}
\arguments{
\item{data}{Matrix of raw counts}
}
\value{
Matrix of transformed data
}
\description{
Apply the log transformation to raw counts
}

