% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Make Heatmap}
\usage{
makeHeatmap(data)
}
\arguments{
\item{data}{Matrix of data to plot}
}
\value{
list of components of the cluster heatmap
}
\description{
Produce the group of plots required for a heatmap with cluster dendrograms
from a matrix
}

