% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_mds.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot MDS}
\usage{
plotMDS(data, top = nrow(data), groups = colnames(data), plot = TRUE)
}
\arguments{
\item{data}{Matrix of data to plot}

\item{top}{Number of rows with highest deviance to select for plotting}

\item{groups}{Vector of groups assigned to sample columns}

\item{plot}{Boolean, if true return plot, if false return plot data}
}
\value{
ggplot2 object containing the PCA plot, or the dataframe of plot data
}
\description{
Produce a MDS plot from a matrix using ggplot2
}

