% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_mds.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA}
\usage{
plotPCA(data, top = nrow(data), groups = colnames(data))
}
\arguments{
\item{data}{Matrix of data to plot}

\item{top}{Number of rows with highes variance to select for plotting}

\item{groups}{Vector of groups assigned to sample columns}
}
\value{
ggplot2 object containg the PCA plot
}
\description{
Produce a PCA plot from a matrix using ggplot2
}

