% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reshape.R
\name{regulariseResults}
\alias{regulariseResults}
\title{Regularise Results}
\usage{
regulariseResults(results, method = c("edgeR", "DESeq", "DESeq2", "voom"))
}
\arguments{
\item{results}{Differential Expression results object}

\item{Method}{Method used to produce the results}
}
\value{
data.frame containing regularised results
}
\description{
Convert results from DE testing to a regular format
}

