% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_isoform_abundance.R
\name{relative_isoform_abundance}
\alias{relative_isoform_abundance}
\title{Quantify relative isoform abundance}
\usage{
relative_isoform_abundance(
  isoform_expression_table,
  isoform_grl,
  ties.method = c("average", "first", "last", "random", "max", "min")
)
}
\arguments{
\item{isoform_expression_table}{A table with the expression values of gene isoforms.
Row names should be the names of the isoforms.}

\item{isoform_grl}{A GRangesList object where the name of each component GRanges
is the name of a gene and the name of the individual regions are the isoform names.}

\item{ties.method}{A character string specifying how ties are treated when ranking isoform expression.
Same as the options for base::rank. Default behavior is to use the mean of tied ranks.}
}
\value{
A RangedSummarizedExperiment with the relative abundance and expression rank of isoforms.
}
\description{
Quantify relative isoform abundance
}
