% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma_plots.R
\name{resultsMA}
\alias{resultsMA}
\title{Results MA}
\usage{
resultsMA(results, method = c("edgeR", "DESeq", "DESeq2", "voom", "regular"),
  alpha = 0.05)
}
\arguments{
\item{results}{Results to plot}

\item{method}{Method used to produce the results}

\item{alpha}{Significance level for labelling differentially
               expressed genes}
}
\value{
ggplot2 object containg MA plot
}
\description{
Produce an MA plot from results of a differential expression test
}

