% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{rlogTransform}
\alias{rlogTransform}
\title{Regularised log transform counts}
\usage{
rlogTransform(data)
}
\arguments{
\item{data}{Matrix of raw counts}
}
\value{
Matrix of transformed data
}
\description{
Apply DESeq2's regularised log transformation to raw counts
}

