% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonR.R
\name{salmon_quantify}
\alias{salmon_quantify}
\title{Run salmon on a group of FASTQ files}
\usage{
salmon_quantify(
  path_to_salmon,
  salmon_index,
  forward_fastqs,
  reverse_fastqs,
  sample_names,
  output_directory,
  merged_output_prefix = "salmon_transcript",
  messages_file = "",
  n_cores = 1
)
}
\arguments{
\item{path_to_salmon}{Path to Salmon executable}

\item{salmon_index}{Path to a salmon index}

\item{forward_fastqs}{A vector with the paths to forward FASTQ files. Each file should be the match of the corresponding file in reverse_fastqs.}

\item{reverse_fastqs}{A vector with the paths to reverse FASTQ files. Each file should be the match of the corresponding file in forward_fastqs.}

\item{sample_names}{A vector with the names of samples corresponding for each pair of samples from forward_fastqs and reverse_fastqs}

\item{output_directory}{The name of the directory to save results in. Will be created if it doesn't already exist.}

\item{merged_output_prefix}{Prefix to use for names of merged output files for counts and TPM which take the form
merged_output_prefix_counts_merged.tsv.gz and merged_output_prefix_tpm_merged.tsv.gz. Default prefix is "salmon_transcript" i.e. default output merged output files are
salmon_transcript_counts_merged.tsv.gz and salmon_transcript_tpm_merged.tsv.gz.}

\item{messages_file}{Name of file to save Salmon run messages. If no file name given, information is printed to stdout.}

\item{n_cores}{The number of cores to use. Default is 1.}

\item{merged_files}{The names of the files in output_directory to save the count and TPM values for all samples to. Defaults are salmon_transcript_count_merged.tsv and salmon_transcript_tpm_merged.tsv}
}
\description{
Run salmon on a group of FASTQ files
}
