% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_mds.R
\name{sampleMDS}
\alias{sampleMDS}
\title{Sample MDS plot}
\usage{
sampleMDS(data, top = nrow(data), groups = colnames(data), plot = TRUE,
  selection = c("pairwise", "common"))
}
\arguments{
\item{data}{Matrix of data to plot}

\item{top}{Number of rows with highest deviance to select for plotting}

\item{groups}{Vector of groups assigned to sample columns}

\item{plot}{Boolean, if true return plot, if false return plot data}

\item{selection}{Select rows in a "pariwise" manner between samples or
                 "common" across all samples}
}
\value{
ggplot2 object containing the PCA plot, or the dataframe of plot data
}
\description{
Produce a MDS plot from a matrix using ggplot2. Based on the limma function.
}

