% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{samtools_fastq}
\alias{samtools_fastq}
\title{Extract forward and reverse reads from BAM files and save to FASTQ files.}
\usage{
samtools_fastq(
  path_to_samtools,
  bam_files,
  fastq_file_prefixes = NULL,
  output_directory = ".",
  parallel_cores = 1,
  compression = ".gz"
)
}
\arguments{
\item{path_to_samtools}{Path to samtools directory}

\item{bam_files}{A vector of paths to BAM files}

\item{fastq_file_prefixes}{Prefixes to use for output FASTQ files.
"_f1.fq" is added to the prefixes for forward reads and "_r2.fq" is added for reverse reads. Default is the basenames of bam_files.}

\item{output_directory}{Path to output directory to save files FASTQ files. Directory is created if it doesn't exist.  Default is current working directory.}

\item{parallel_cores}{The number of files to process at a time in parallel. Default is just 1}

\item{compression}{One of ".gz" to compress files with gzip, ".bgzf" to compress files with bgzip or NULL to not compress files. Default is gzip compression.}
}
\description{
Took 4.5 minutes with a BAM of size 2.8 GB.
}
