% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/foldchange.R
\name{setTable}
\alias{setTable}
\title{Gene Set Table}
\usage{
setTable(data.list, de.set, gene.sets)
}
\arguments{
\item{data.list}{List of differential expression results}

\item{de.set}{Set of differentially expressed genes}

\item{gene.sets}{Named list of gene sets to include in table}
}
\value{
data.frame containing the summary table
}
\description{
Produce a table showing how many genes in a list of sets are differentially
expressed, up-regulated and down-regulated.
}

