% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{showHeatmap}
\alias{showHeatmap}
\title{Show heatmap}
\usage{
showHeatmap(plot.list, col.width = 0.2, row.width = 0.2)
}
\arguments{
\item{plot.list}{List of plot components returned by makeHeatmap}

\item{col.width}{Width of the column cluster dendrogram}

\item{row.width}{Width of the row cluster dendrogram}
}
\description{
Assemble and display the components of a cluster heatmap
}

