% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{sort_bams}
\alias{sort_bams}
\title{Sort BAM files}
\usage{
sort_bams(
  path_to_samtools,
  bam_files,
  sorted_bam_files = NULL,
  parallel_cores = 1
)
}
\arguments{
\item{path_to_samtools}{Path to samtools directory}

\item{bam_files}{Paths to BAM files}

\item{sorted_bam_files}{Paths to save sorted BAM files. Note that sortBam adds .bam to end of file even if it is already present.
Default is to add "sorted" to end of names of bam_files}

\item{parallel_cores}{Number of cores to use. Default is 1.}
}
\description{
Sort BAM files
}
