% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sratools.R
\name{sra_fastq_dump}
\alias{sra_fastq_dump}
\title{Download SRA files with prefetch}
\usage{
sra_fastq_dump(
  path_to_sratk,
  srr_directory_list,
  output_directory = ".",
  parallel_files = 1,
  compress_fastq_files = T,
  remove_srr_directories = F
)
}
\arguments{
\item{path_to_sratk}{Path to SRA toolkit bin directory}

\item{srr_directory_list}{A vector of sequence read accessions}

\item{output_directory}{Path to output directory to save files. Default is current working directory.}

\item{remove_srr_directories}{A logical value indicating whether SRR directories should be removed after FASTQ files have been extracted. Default is FALSE.}

\item{parallel_download}{The number of files to download at a time in parallel. Default is just 1}
}
\description{
Download SRA files with prefetch
}
