% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sratools.R
\name{sra_prefetch}
\alias{sra_prefetch}
\title{Download SRA files with prefetch}
\usage{
sra_prefetch(
  path_to_sratk,
  srr_accessions,
  output_directory = ".",
  parallel_files = 1,
  ngc_key = NULL
)
}
\arguments{
\item{path_to_sratk}{Path to SRA toolkit bin directory}

\item{srr_accessions}{A vector of sequence read accessions}

\item{output_directory}{Path to output directory to save files. Default is current working directory.}

\item{ngc_key}{Path to an ngc repository key.}

\item{parallel_download}{The number of files to download at a time in parallel. Default is just 1}
}
\description{
Download SRA files with prefetch
}
