% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{transformCounts}
\alias{transformCounts}
\title{Transform counts}
\usage{
transformCounts(data, methods, verbose = TRUE)
}
\arguments{
\item{data}{Matrix of raw counts}

\item{methods}{Vector of transformation methods to apply, choices are "log",
"vst", "rlog", "cpm" and "logCPM".}

\item{verbose}{Boolean, whether to print messages showing progress}
}
\value{
List of transformed matrices
}
\description{
Apply transformations to raw counts for applications such as plotting
}

