% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{vennDiagram}
\alias{vennDiagram}
\title{Venn Diagram}
\usage{
vennDiagram(data.list, alpha = 0.05)
}
\arguments{
\item{data.list}{List of results to combine}

\item{alpha}{Significance level for selecting genes}
}
\value{
Venn diagram object
}
\description{
Produce a Venn diagram from a list of differential expression results
}

