% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{vennGenes}
\alias{vennGenes}
\title{Venn Genes}
\usage{
vennGenes(data.list, alpha = 0.05)
}
\arguments{
\item{data.list}{List of results to combine}

\item{alpha}{Significance level for selecting genes}
}
\value{
List of vectors of genes
}
\description{
Get list of genes for each region in a Venn diagram from a list of
differential expression results.
}

