% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(results, method = c("edgeR", "DESeq", "DESeq2", "voom",
  "regular"))
}
\arguments{
\item{results}{Results to plot}

\item{method}{Method used to produce the results}
}
\value{
ggplot2 object containg volcano plot
}
\description{
Produce a volcano plot from differential expression results
}

