% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalise.R
\name{voomNorm}
\alias{voomNorm}
\title{voom Normalise}
\usage{
voomNorm(dge)
}
\arguments{
\item{dge}{DGEList object to normalise}
}
\value{
Normalised EList object
}
\description{
Normalise a DGEList object using limma-voom
}

