% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{voomTest}
\alias{voomTest}
\title{voom Test}
\usage{
voomTest(voom.data, group1, group2)
}
\arguments{
\item{voom.data}{EList object to test}

\item{group1}{First group to test, the reference or control}

\item{group2}{Second group to test, the treatment}
}
\value{
dataframe containing test results
}
\description{
Test a normalised EList object using limma-voom
}

