% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{vstTransform}
\alias{vstTransform}
\title{VST transform counts}
\usage{
vstTransform(data)
}
\arguments{
\item{data}{Matrix of raw counts}
}
\value{
Matrix of transformed data
}
\description{
Apply DESeq's variance stabilising transformation to raw counts
}

