% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{edgeRTest}
\alias{edgeRTest}
\title{edgeR Test}
\usage{
edgeRTest(dge, group1, group2)
}
\arguments{
\item{dge}{Normalised DGEList object to test}

\item{group1}{First group to test, the reference or control}

\item{group2}{Second group to test, the treatment}
}
\value{
dataframe containing test results
}
\description{
Test a normalised DGEList object using edgeR
}

