% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{listVolcano}
\alias{listVolcano}
\title{List Volcano}
\usage{
listVolcano(data.list)
}
\arguments{
\item{data.list}{List of results to plot}
}
\value{
List of ggplot2 object containing volcano plots
}
\description{
Plot volcano plots from a list of differential expression results
}

