% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{transformCounts}
\alias{transformCounts}
\title{Transform counts}
\usage{
transformCounts(data, methods = c("log", "vst", "rlog", "cpm"))
}
\arguments{
\item{data}{Matrix of raw counts}

\item{methods}{Vector of transformation methods to apply, choices are "log",
               "vst", "rlog" and "cpm".}
}
\value{
List of transformed matrices
}
\description{
Apply transformations to rawcounts for applications such as plotting
}

